Add support for a synchronous delete, via observe operation.
This change adds support for a synchronous delete, including
detection of full delete versus logical delete.

The main use case behind this feature is that if using
Couchbase Server 2.0, and wanting to query views, one may
wish to ensure a delete has been persisted to know that
a subsequent query with .setStale(false) will take the
deletion into account when generating the index.

Change-Id: I99ad5b5089992a7df2c2ab12e487d043265df21f
Reviewed-on: http://review.couchbase.org/19778
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>