JCBC-96: Remove the queue mechanism in ViewConnection
This change removes the queueing mechanism that hold operations
before being given to an individual ViewNode. The operations are
now passed directly to a ViewNode and if there are too many
operations then they are held as pending requests until the
ViewNode is ready to process them. On node shutdown the pending
ops are requeued through a callback function that redistributes
them equally to the rest of the ViewNodes.

Change-Id: Ieffbb45123cf47153cea662df1a01f5b2266c8d3
Reviewed-on: http://review.couchbase.org/19606
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>