Change tests to use bucket delete/create instead of flush_all.
Recent changes of Couchbase Server 2.0 have flush_all disabled
by default. It's also not correct to re-enable flush_all
(though possible) as it won't really behave correctly.

This change shifts to a bucket delete followed by a bucket create.
It's not the best change in that it introduces a lot of sleep time
instead of checking for bucket availability with some other poll,
but does appear to be reliable.

After the integration of RESTful flush in MB-5170, we should be
able to trim some of the time of running tests again and remove
the sleeps in here.

Change-Id: I80ed8bb311afc8dde3970db0cbad6687f58c0048
Reviewed-on: http://review.couchbase.org/19677
Tested-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>