JCBC-109: Reduce observe poll interval latency.
A 100ms interval was chosen after some discussion.  It's
not optimal, and JCBC-108 tracks getting this to an adaptive
algorithm.

The reason for 100ms is that with real disks, chances are
fast IOs will happen in 60-120ms.  Only with caching or SSDs
will it likely be lower.  This means that with most checks
it'll happen in a handful of checks.

Change-Id: Ia6e14ee32ba6d6fc57608ffb44dbb9611e3401a5
Reviewed-on: http://review.couchbase.org/20760
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>