JCBC-41: Allows better query options through ComplexKey.
Related Issues: JCBC-90, JCBC-48

This changeset provides more flexible options on how to provide
the query params, especially if someone wants to add numeric
strings and not want them to be casted to pure numbers. Since we
only allow Strings or ComplexKeys as params, the ComplexKey class
can be used to hint the correct type and enfore it down to the
HTTP query layer. See the added test cases for more details.

Change-Id: I0d56bda1d3d9727455bfe26fa7a1311dbcce00a8
Reviewed-on: http://review.couchbase.org/21337
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>