JCBC-120: Fix NPE on null for password
This changeset sets the password to an empty string if it is passed
in as null. This fixes a NPE exception down the stack in the spy
memcached library that tries to access it. This of course also sets
it for couchbase connections, but since an empty string is always
allowd it makes it more consistent.

Change-Id: Ic62558d2dbbef8e67f7570eb605207587718aec7
Reviewed-on: http://review.couchbase.org/21446
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>