JCBC-130: Add a timeout for unresponsive http connection fetches.
This adds a hardcoded timeout and a test to make sure that after
a certain time it is guaranteed that the thread waiting for the
connection initialization returns eventually.

The timeout is set to 30 seconds. Note that this is not the
timeout for the actual view response to receive, but for the
connection initialization.

Change-Id: Icb1138803d7914853b4e0f69b5a8a0443e4170c0
Reviewed-on: http://review.couchbase.org/22206
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>