JCBC-123: Throw an exception when vbucket master is -1
This changeset makes sure to throw a controlled exception when
vor the given vbucket there is no master server defined to handle
it appropriately. -1 for the master is clearly an invalid state
and could be the case for the following reason:

- No replicas are defined and a node is failed over.
- More nodes have been failed over than there are replicas.

Either way, the client library is unable to remedy the situation
on its on and therefore throws an exception. The application
layer is expected to deal with the situation (either retry until
a rebalance is done) or redirect it to a system which is capable
to do so.

Change-Id: I582939820ec3067ce724c93e410e93f834c340ee
Reviewed-on: http://review.couchbase.org/22352
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>