JCBC-135: better interoperability between the client and netty.
This changeset make the netty channel wait for inside a listener, so
that it possible to use the couchbase client inside a netty-based
server architecture. Netty by default forbids the waiting inside
a worker, therefore we need to assign the channel on a different
thread and then wait based on a Latch.

Change-Id: Ic14d0ec5dcd39e16986aa1c3386f0320bd386b41
Reviewed-on: http://review.couchbase.org/22589
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Deepti Dawar <deepti.dawar@globallogic.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>