JCBC-148: check persistence against node count
This changeset verifies the persistence requirements against the
total node count and returns an exception if the requirements can
not be satisfied by the current cluster layout.

Note that this does NOT look at how much nodes are available for
a given key, this needs to be implemented as an enhancement.
The corresponding enhancement can be found at JCBC-161.

Change-Id: I531ad01b332a367256a8a1c6a0941b3683dede9e
Reviewed-on: http://review.couchbase.org/22936
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>