JCBC-167: Allow long values for ComplexKeys.
This changeset prevents the bug of throwing quotes around
long values. Now it is possible to use long values through
the ComplexKey object. Since this is tried as a last resort,
no old behavior is subject to change.

This also adds some additional tests around booleans and
null values since those are primitives to JavaScript.

Change-Id: I60c68aa48631b719cc7b5ec0328b257380a64874
Reviewed-on: http://review.couchbase.org/23194
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>