JCBC-214: Add read timeout on configuration fetch
This changeset adds a 5 second read timeout when the configuration
is fetched from the server. This prevents a possible issue when
the couchbase server process is hanging, but the socket is still
open. In this case, the client now skips to the next in the given
list of nodes when there is no response after 5 seconds coming
back (even when the socket has been opened properly).

Change-Id: I155b522648d2e4b4848001b03e6eeb443dacad71
Reviewed-on: http://review.couchbase.org/24019
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>