JCBC-219: Fix reconnect logic on valid rebalance out/failover.
This changeset forces to reconnect the CouchbaseClient when a
node is rebalanced out, but still alive. The connection is
correctly closed for netty and not "hard" disconnected. This
prevents an issue where the client is left in a state where
it does not receive new cluster updates because it still
listens on a node that has been removed.

Change-Id: Ideefe229d22ee5becc3c6401d43f773a04aba56f
Reviewed-on: http://review.couchbase.org/24182
Reviewed-by: Deepti Dawar <deepti.dawar@globallogic.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>