JCBC-351: Harden asyncGetFromReplica.
This changeset makes sure that also the active node
is considered in the replica list (to get a value
even after a failover happened in the meantime) and
also make sure that if no replica node is active for
the key, reconfiguration is triggered to avoid dead
nodes.

This is in alignment with the getPrimary() operations
which use the same behavior.

Change-Id: I5a9c7c888025069942c3db366e36dbd627f643e0
Reviewed-on: http://review.couchbase.org/28473
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>