JCBC-138: Replace nodeList on reconfiguration to reflect changes.
This changeset makes sure that, after a reconfiguration, the new
node list is passed on to the configuration handler so that it is
aware of the new topology.

This changeset now makes it possible that also when only one node
is passed in during construction, all of the other nodes will be
picked up proactively.

Change-Id: I810f6b4e419a1ddecf58c46d428ada66e17a84c2
Reviewed-on: http://review.couchbase.org/28785
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>