JCBC-369: Fix observePoll delete logic.
This changeset refactors the observePoll method, making it faster
and also correcting the delete logic which is not correct for
certain scenarios (now it is correct with respect to the non-delete
methods).

Also, it optimizes the observe() method to only add those nodes
to the bcast list that are actually needed to fulfill the
constraint.

Change-Id: Ib101af5f1212123fcb103d335d40fb2bc79f8964
Reviewed-on: http://review.couchbase.org/29545
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>