JCBC-374: Also trigger completion on active node replica doc.
This changeset makes sure that when a replica get (that is
fanned out to all replica nodes + the active node) from the
active node returns first, the replica get future is also
completed.

This fixes the situation where 1 replica is defined, this node
is failed over and no data is returned (while of course the
master node is still reachable and returns the result). Before
the data was returned properly but the latch on the replica get
future was never counted down, and as a result never indicating
success.

Change-Id: Ib6b513819ed9a686fdc80956ebb58edf9cfc6a43
Reviewed-on: http://review.couchbase.org/29848
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>