JCBC-361: Add specific observe timeout.
This changeset adds a longer timeout for futures when persistence
or replication constraints (observe) are used. It currently defaults
to 5 seconds, acknowledging that writing to disk may take longer
than just writing to ram.

The manual setting of maximum number of polls has been deprecated,
because it is now calculated out of the observe timeout and the
observe interval, making it more flexbile and easier to configure.

In summary, regular operations still have the 2.5 second timeout,
but everywhere "PersistTo" or "ReplicateTo" is used, a longer,
5 second timeout is used instead by default.

Change-Id: Ifdee7550782c72bb71bc11ddede2c139e3b5436e
Reviewed-on: http://review.couchbase.org/31655
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>