Harden "bucket state"-dependent test cases.
This changeset makes sure that various test cases which depend
on bucket state (that is mostly warmup happening) are more
resilient to false-negatives. This is mostly achieved through
larger allowed timeouts, which allows slower disks also to
catch up between the unit tests.

It increases the timeout for the http cluster
operations in the test from 2.5 seconds (operationTimeout) to
the more realistic HttpTimeout (75 seconds).

Change-Id: I4d884c74802ad9906b3a44e91ff9b45f1ca6b650
Reviewed-on: http://review.couchbase.org/32306
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>