JCBC-427: Update enclosed HTTP provider with new configs.
This changeset makes sure that the enclosed HTTP configuration
provider also gets new configurations during runtime, so that once
it needs it, it doesnt override a good one with a very old one.

This is particularly important when the HTTP connection gets closed,
because then a "notifyDisconnect" would override a good config with
a old and outdated one, leading to inconsistencies.

Change-Id: Iffce8028c49844eb5a834fcfc1e93e7525128536
Reviewed-on: http://review.couchbase.org/34378
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>