JCBC-436: Expose custom auth wait time setting.
Motiviation
-----------
With changes introduced in spy (SPY-159), it is possible to customize
the auth wait time on the connection factory builder. This time should
be increased for the couchbase SDK and also still be configurable.

Modifications
-------------
The setting is exposed through the regular methods, test have been added.
Note that the default time has been increased to 2.5 seconds, which mirrors
the default OP timeout.

Slight-side effect change: one test has been adapted to some interface
changes in the spymemcached library, which needs to be part of this changeset.

Result
------
The auth wait timeout is now configurable.

Change-Id: I7d2acc94e2fbd220154caf8c5dfe787b9374ca5d
Reviewed-on: http://review.couchbase.org/34854
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>