JCBC-439: Fix AuthDescriptor overriding.
Motivation
----------
The AuthDescriptor, even when set through the builder, is ignored on factory
construction. This means that no manual descriptor can be passed in, always
leading to SASL auth list.

Modifications
-------------
The appropriate methods in the builder have been implemented and tests added
to verify the functionality.

Result
------
A custom descriptor can now be passed in and be used to override default auth
settings.

Change-Id: I66ac4e34a32e2da31d1b8a784bcfb39e9ce31f01
Reviewed-on: http://review.couchbase.org/34935
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>