Make sure CCCPAwarenessTest always has a bucket.
Motivation
----------
It could be the case that the CCCPAwarenessTest does not have a bucket
to use if used in the wrong order by JUnit for all tests.

Modifications
-------------
The fix makes sure that a default bucket is created, it also makes sure
that the test doesn't fail against more than one node incorrectly.

Result
------
The test is now more failure resilient to false positives.

Change-Id: I617561d900129e8ddda86e6ad624760473258a5f
Reviewed-on: http://review.couchbase.org/35722
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>