JCBC-453: Also check for fresh configs with failure modes != redistribute
Motivation
----------
When a different failure mode is used, it can be the case that a new
configuration is not picked up (ie with cancel).

Modification
------------
This changeset also makes sure we update the config reload threshold
if other failure modes are used.

Result
------
New configurations eventually get picked up even if other failure
modes are used.

Change-Id: I302b2192f2fc34ea7df36d3b83ce63197513d7be
Reviewed-on: http://review.couchbase.org/36828
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>