JCBC-457: Force CCCP config fetching on node reconnect.
Motivation
----------
If a node needs to be reconnected, there is a strong indication that
the socket has been closed and this could be because of a topology
change.

Modification
------------
If a reconnect is scheduled, make sure it forces a config update.
Also, the method is added for memcache buckets to keep the behavior consistent.

Result
------
Quicker detection of topology changes, eventually getting quicker to
a valid config state.

Change-Id: I5244dfc6d6f19288977ef98745d47efe25773093
Reviewed-on: http://review.couchbase.org/36779
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>