JCBC-459: Ignore Configurations with old (or same) revID.
Motivation
----------
If carrier config is used, new configuration arrives in the payload of the
NMVB responses. Since configurations trickle through the server cluster not
100% at the same time, it can happen that we get a outdated configuration from
one of the nodes, where the client already is in the posession of a newer
one.

The server emits a revid with a monotonically increasing number, so if we
have a higher revid we can skip/discard incoming older ones and keep going.
By not ignoring them, the client may run into a situation where it updates
a older config in place for a newer one.

Modifications
-------------
The revnr is parsed out of the configuration, but if no id is set (for backwards
compatibility reasons), -1 is used. In the provider itself, if the number is > 0
it is checked against the current known revnr and if higher the new config
is used and the subscribers are notified.

Result
------
Older configurations will now not be used anymore if a newer version is already
in the posession of the SDK.

Change-Id: I0eb5af448e8149017962030de6fe054358bd67ea
Reviewed-on: http://review.couchbase.org/36981
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>