JCBC-460: Make sure replica vbucket changes count as significant.
Motivation
----------
When a new configuration arise, there is a simple difference heuristic in place
which checks if a new configuration is worth applying or if it only contains
insignificant changes that are not worth the effort of applying a new configuration.

One of those measurements is if vbuckets have changed (the values in the vbuckets
array). One case that was not thought of before is if only replica vbuckets change
and the master buckets stay the same. This has been recently found to be an issue
when configurations are not applied because only the replica vbucket information
has changed.

Modifications
-------------
The algorithm now also checks the content of the replica vbuckets to see if
a significant modification has been made, leading to an update of the map
to the according parties.

Result
------
New configuration is now also applied if only the replica vbuckets change. A
test case has been added to verivy the functionality (and fails without the
change).

Change-Id: I50d81f0053360b1c073fa90c217bde18abafd78d
Reviewed-on: http://review.couchbase.org/36985
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>