JCBC-413: Fix race condition in ClusterManager.
Motivation
----------
When the ClusterManager response returns with a successful response, it could be that the code waiting
on the result is notified before the corresponding results are stored.

Modifications
-------------
The latch is now counted down after the results have been gathered, avoidin the race condition.

Result
------
The code afterwards waiting on the latch is now only counted down before the results have been gathered.

Change-Id: I97f0f0eb72e74a6400fecd5d03e8d35e537b4339
Reviewed-on: http://review.couchbase.org/37719
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>