JCBC-424: Broadcast NOOP on idle state.
Motivation
----------
This changeset has two main motivations:

  - Detect rebalancing out nodes  & broken sockets even when there is no load.
  - Help firewalls that close idle sockets.

Modifications
-------------
In addition to a change in spymemcached which calls a method if the selector is woken up
without actually selecting something (so the delay hits), this change correlates the last
write with the woken up time. If the timespan is longer than a given maximum, a NOOP
broadcast is sent, which works like any other operation but immediately triggers broken
sockets and therefore a new configuration will be fetched.

Note that the NOOP broadcast is not sent under load, since the write timestamp will always
be fresh (and the more load the more seldom the method itself will be called from spy).

Result
------
Rebalancing away nodes and broken sockets are much quicker detected and also the socket is
prevented from being closed becasue it is idling too much in case of no load.

Change-Id: I056b76ea009a689f4ba700730fb281d78d57484d
Reviewed-on: http://review.couchbase.org/37729
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>