JCBC-476: Rework misleading log message.
Motivation
----------
The message in the changeset is misleading because the ops are now
queued up and therefore more resilient. Also, WARN is too high since
it should be just an INFO message. If auth or anything else fails,
different messages will be shown in the logs anyway.

Modifications
-------------
Change it to INFO level and also rework the text to be more clear
what is happening with the given key.

Result
------
No misleading information anymore which looks falsly suspicious.

Change-Id: I9b0052dd82bad4edbc54df4d06c98b6b18f31dce
Reviewed-on: http://review.couchbase.org/38668
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>