JCBC-477: Pass down the timeout to the getBulk for views
Motivation
----------
If no timeout is passed down to the BulkGetFuture, Long.MAX_VALUE
is used as a timeout which, if something goes wrong, basically
translates to block the thread forever. Also, this breaks the
contract since a custom time can always be passed in.

Modifications
-------------
The timeout is now passed down to the bulk get so that a timeout
is properly respected.

Result
------
Correct behavior from the client side.

Change-Id: I56696a21ed14e9c0669eb5d05941f67a987eeb18
Reviewed-on: http://review.couchbase.org/38678
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>