JCBC-480: Complete replica read ops even if NOT_EXISTS
Motivation
----------
get(s)FromReplica operations may timeout incorrectly if the document has not been
stored on the server at all, since it waits for a successful response before it
completes the latch and stores the response. Since a non-existing doc response
returns NOT_EXISTS and false success, it will never complete and timeout.

Modifications
-------------
Complete the latch as well if the response is a NOT_EXISTS, indicating a good
response but the document was just not found. This is perfectly fine under the
assumptions that the response is allowed to be non-consistent by contract, even
if it may not be replicated yet.

Result
------
The behavior is now consistent for both found and not found docs.

Change-Id: Ic310255e28c498ed6482101f64657978fad8c7ae
Reviewed-on: http://review.couchbase.org/38796
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>