JCBC-482: Select proper replica node for getsFromReplica
Motivation
----------
getsFromReplica did not work when the master node is currently failing because
it had not been scheduled properly. This fix remedies that. Note that
getFromReplica had always worked as expected.

Modification
------------
This changeset adds support for getsFromReplica in the add operation codepath
to properly select the replica node.

Result
------
getsFromReplica now works like getFromReplica even when only replica nodes
are available.

Change-Id: I4362efbba4ea04ff063c85128949a3931c95bd58
Reviewed-on: http://review.couchbase.org/38866
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>