JCBC-464: Clear buckets for HTTP provider on (re)bootstrap.
Motivation
----------
Prior to this change, a combination of (re)bootstrapping initiated
by a full cluster shutdown and a previously loaded configuration
led to a invalid state in the HTTP provider. Since it is also used
for a cccp fallback, it always interfered in such a scenario.

Modifications
-------------
The internal problem was that once bootstrapped, the bucket config
was stored and on a re-bootstrap, this old config was taken. That
itself did not cause trouble, but the http streaming config attachment
was broken because some parameters were not set properly (which would
have been on a full bootstrap).

This changeset clears out the bucket configs before a (re)bootstrap
to make sure a full http walk cycle is always done. This provides
more predictable behavior and also avoids a reported NPE.

Result
------
Even if rebootstrapped, it now either gets to a valid configuration
over HTTP and succeed or will retry, without failing with a NPE or
ending up in an invalid state.

Change-Id: Id8644b1fddf7b38168e663a3d5af51e17e56b9c4
Reviewed-on: http://review.couchbase.org/39346
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>