JCBC-488: Make sure headers are overriden, not appended on readd.
Motivation
----------
If a view operation needs to be retried, the current code adds headers
instead of overwriting them, which leads to broken http requests. As a
result, the server responds with 400 and is unable to process the query.

Modifications
-------------
Changing the code from addHeader to setHeaders overrides the headers
instead of appending them.

Result
------
Valid HTTP requests, even if it needs to be retried.

Change-Id: I13b72effc2998d17aaf2ac8acfc3c225aef30de3
Reviewed-on: http://review.couchbase.org/39429
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>