JCBC-575: Add support for a RawJsonDocument.
Motivation
----------
This changeset adds the capabilities to pass in a raw (already encoded)
JSON document, largely driven by the fact that users may already have their
JSON stack set up and just want to pass in the data. We do not want to get
in the way.

Modifications
-------------
Adds the raw JSON document which justs takes the string, expects its valid,
creates a buffer out of it and stores it down. This also is very low overhead.

Result
------
More flexibility in dealing with JSON documents in combination with custom
JSON marshallers.

Change-Id: Ibd7e37ae48de197df72713acd7f4daeadc352904
Reviewed-on: http://review.couchbase.org/41945
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>