JCBC-578: Validate bucket name before connecting.
Motivation
----------
The bucket name is normally not null when added by the developer at compile
time, but in environments like spring boot or others with runtime configurations
it can happen that the bucket name is null for some reason.

Modifications
-------------
The code now properly checks for that and throws a descriptive exception instead
of failing with an NPE down the codepath.

Result
------
It is immediately clear what went wrong on connecting.

Change-Id: I0bbcdd049b7e9888d467084da29560b0cfa35879
Reviewed-on: http://review.couchbase.org/41986
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>